split out counterparty types
